% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armadillo-package.R
\docType{package}
\name{armadillo-package}
\alias{armadillo}
\alias{armadillo-package}
\title{armadillo: An 'Armadillo' Interface}
\description{
Provides function declarations and inline function definitions that facilitate communication between R and the 'Armadillo' 'C++' library for linear algebra and scientific computing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/armadillo/}
  \item \url{https://github.com/pachadotdev/armadillo}
  \item Report bugs at \url{https://github.com/pachadotdev/armadillo/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
\keyword{internal}
