% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{armadillo_version}
\alias{armadillo_version}
\title{Get Armadillo version}
\usage{
armadillo_version()
}
\value{
A string with the Armadillo version name and number
}
\description{
Provides the Armadillo C++ library version name and number included in the
package.
}
\examples{
armadillo_version()
}
