% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11armadillo-package.R
\docType{package}
\name{cpp11armadillo}
\alias{cpp11armadillo-package}
\alias{cpp11armadillo}
\title{cpp11armadillo: A template for an R package with cpp11 and armadillo}
\description{
Add some description later. This is about linear algebra.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pacha.dev/cpp11armadillo/}
  \item \url{https://github.com/pachadotdev/cpp11armadillo}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11armadillo/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

}
