% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-template.R
\name{pkg_template}
\alias{pkg_template}
\title{Start a new project with the cpp11armadillo package template}
\usage{
pkg_template(pkgname = NULL, path = NULL)
}
\arguments{
\item{pkgname}{Name of the new package (defaults to the name of the current
working directory)}

\item{path}{Path to the new project (defaults to working directory)}
}
\description{
Start a new project with the cpp11armadillo package template
}
